adding types to model directory
