+++ credentials/src/main/kotlin/model/CredentialStatus.kt
@@ -1,0 +1,3 @@
+package web5.credentials.model
+
+public typealias CredentialStatus = com.danubetech.verifiablecredentials.credentialstatus.CredentialStatus
