+++ credentials/src/main/kotlin/model/CredentialSubject.kt
@@ -1,0 +1,3 @@
+package web5.credentials.model
+
+public typealias CredentialSubject = com.danubetech.verifiablecredentials.CredentialSubject
