+++ credentials/src/main/kotlin/model/FieldV2.kt
@@ -1,0 +1,8 @@
+package web5.credentials.model
+
+public data class FieldV2(
+  val id: String?,
+  val path: List<String>?,
+  val purpose: String?,
+  val name: String?
+)
