+++ credentials/src/main/kotlin/model/Format.kt
@@ -1,0 +1,7 @@
+package web5.credentials.model
+
+public data class Format(
+  val jwt: JwtObject?,
+  val jwtVc: JwtObject?,
+  val jwtVp: JwtObject?
+)
