+++ credentials/src/main/kotlin/model/JwtObject.kt
@@ -1,0 +1,5 @@
+package web5.credentials.model
+
+public data class JwtObject(
+  val alg: List<String>,
+)
