+++ credentials/src/main/kotlin/model/VerifiableCredentialType.kt
@@ -1,0 +1,3 @@
+package web5.credentials.model
+
+public typealias VerifiableCredentialType = com.danubetech.verifiablecredentials.VerifiableCredential
