+++ credentials/src/main/kotlin/model/VerifiablePresentationType.kt
@@ -1,0 +1,3 @@
+package web5.credentials.model
+
+public typealias VerifiablePresentationType = com.danubetech.verifiablecredentials.VerifiablePresentation
