+++ credentials/src/test/kotlin/SSITest.kt
@@ -6,6 +6,8 @@ import org.junit.jupiter.api.BeforeEach
 import uniresolver.result.ResolveDataModelResult
 import uniresolver.result.ResolveRepresentationResult
 import uniresolver.w3c.DIDResolver
+import web5.credentials.model.CredentialSubject
+import web5.credentials.model.VerifiableCredentialType
 import java.net.URI
 import java.util.Base64
 import java.util.Date
